+++ Runtime/CreateAssetMenuConstants.cs
@@ -24,6 +24,7 @@ namespace Zor.RandomGenerators
 		public const string UnityDistributionFolder = UniformDistributionsFolder + "Unity Distribution/";
 
 		public const string DiscreteDistributionsFolder = RandomGeneratorProvidersFolder + "Discrete Distributions/";
+		public const string BernoulliDistributionFolder = DiscreteDistributionsFolder + "Bernoulli Distribution/";
 		public const string PoissonDistributionFolder = DiscreteDistributionsFolder + "Poisson Distribution/";
 		public const string WeightedDistributionFolder = DiscreteDistributionsFolder + "Weighted Distribution/";