+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Random generator providers infrastructure.
 - Acceptance-Rejection distribution.
 - Bates distribution.
+- Bernoulli distribution.
 - Box-Muller distribution.
 - Irwin-Hall distribution.
 - Marsaglia distribution.